requirements.txt
streamlit==1.31.0
requests==2.31.0
torch==2.2.0
transformers==4.37.2
Pillow==10.2.0
matplotlib==3.8.2
numpy==1.26.3
google-generativeai==0.3.2
python-dotenv==1.0.1